neon_local init: properly stop & wait for pageserver child process to exit
Before this patch, we would SIGKILL but not wait for the child to exit,
but still remove its pidfile.

If a `neon_local start` command follows, that would then successfully spawn
a pageserver process, but the pageserver process would fail to bind the
listener ports because the old child child (from neon_local init) was
still around.
The error returned by the kernel in that case is EADDRNOTAVAIL.
We've encountered it frequently in CI:

    2022-11-29T18:14:38.495970Z  INFO Starting pageserver http handler on localhost:15041
    Error: Failed to start pageserver

    Caused by:
        Cannot assign requested address (os error 99)

refs https://github.com/neondatabase/neon/issues/2968