Remove bogus ensure! on pid-file removal.
I originally intended to add code to pageserver shutdown, to remove the PID
file, and the intention of this ensure! was to check that the PID file *is*
removed. But then I figured it's too hard to remove the PID file safely in the
pageserver, and decided to just not do it. At the same time, I accidentally
got the ensure! check backwards, so it was actually checking that the PID
file is *not* removed. That's why it was not failing.

The check is quite pointless and confusing as it is, so remove it.