refactor: split off LayerMap's needs from trait Layer into super trait PureLayer
While this complicates the typings in `layer_map.rs`, it removes the
need for most `unimplemented!()` / `todo!()` impls in the layer types
used for benchmarking the layer map.

That by itself wouldn't justify the time I put into this refactoring.

But, after this change, I will remove `PathBuf`s from the `Layer` trait's interface,
which in turn will enable a clean-up of the `storage_sync2` APIs.