Avoid deleting a layer file that is still in use.
Firstly, reorder the calls to remove a Layer from the layer map, and
removing the underlying file, so that the Layer is removed from the
layer map first. Otherwise, another thread could get the reference to
a Layer that refers to a file that we just deleted.

That's not enough to fix the race condition though. It's still
possible that another thread looked up the Layer just before we
removed it from the layer map, and tries to read it after we have
already deleted the underlying file. To close that race, add a new
mechanism to wait for the reference count of the Arc<Layer> to fall to
1, ensuring that there are no other references to it before we delete
it. This race is expected to happen very seldom, and it's not too bad
if compaction or GC has to wait a little longer in the rare cases, so
just use a simple polling with exponential backoff for the waiting.