Timeline::download_missing: fix handling of mismatched layer size
Before this patch, when we decide to rename a layer file to backup
because of layer file size mismatch, we would not remove the layer from
the layer map, but remote the on-disk file.

Because we  re-download the file immediately after, we simply end up
with two layer objects in memory that reference the same file in the
layer map.  So, GetPage() would work fine until one of the layers gets
delete()'d. The other layer's delete() would then fail.