task_mgr: use CancellationToken instead of shutdown_rx
this should help us in the future to have more freedom with spawning
tasks and cancelling things. CancellationToken allows creation of
hierarchical cancellations, which would also simplify the task_mgr
shutdown operation, rendering it unnecessary.

think of a hierarachy where in the root you have:

Tenant
 |- timeline1
 |   \- timeline1 initial size calc
 |- timeline2
 |   \- timeline2 foobar operation
 \- timeline3
     \- yet another operation

All of these need to shutdown when tenant is shutdown, and the
cancellation token hierarchy would support this, allowing phasing out of
the task_mgr.