[2/3] use new batch layer upload scheduling API where possible
This comes with a few redundant IndexUploads in places where the
layer upload is followed by layer deletions.

But the next patch in this series will convert
schedule_layer_file_deletion into a no-op if the list of layers to
delete is empty.
Places that upload layers shouldn't implicitly rely on that