replace RemoteLayer's DownloadState logic with a Sempahore
The previous implementation was quite verbose and had a race
condition - we would need an async condvar to make it work correctly.

As a replacement, use a tokio::Semaphore that we close if the
download was successful.
We use a similar, albeit not identical approach, for initial logical
size computation.