Fix issue in compaction code
If we ran `compact_prefetch_buffers` with exactly one hole in the buffers,
the code would fail to remove the last, now unused, entry from the array.

This is now fixed. 

Also, add and adjust some comments in the compaction code so that
the algorithm used is a bit more clear.