Remove code and test to generate flamegraph on GetPage requests.
It was nice to have and useful at the time, but unfortunately the
method used to gather the profiling data doesn't play nicely with
'async'. PR #3228 will turn 'get_page_at_lsn' function async, which
will break the profiling support. Let's remove it, and re-introduce
some kind of profiling later, using some different method, if we feel
like we need it again.