If an error happens during consumption metrics collection, continue.
Previously, any error on any tenant would cause the whole task to
exit. That's not nice, because then no more metrics will be collected
for any tenant.

With this, if we get stuck on one tenant, we still stop the collection
cycle and don't process the other tenants. That's not great.