make Timeline::{get,get_reconstruct_data} aware of context in which they are called
Motivation
==========

Layer Eviction Needs Context
----------------------------

Before we start implementing layer eviction, we need to collect some
access statistics per layer file or maybe even page.
Part of these statistics should be the initiate of a page read request. I.e.,
was it page_service vs. a background loop, and if the latter, which of them.

Further, it would be nice to learn more about what activity in the pageserver
initated an on-demand download of a layer file.
We will use this information to test out layer eviction policies.

Read more about the current plan for layer eviction here:
https://github.com/neondatabase/neon/issues/2476#issuecomment-1370822104

task_mgr problems + cancellation + tenant/timeline lifecycle
------------------------------------------------------------
Apart from layer eviction, we have long-standing problems with task_mgr,
tenant / timeline deletion, and generally, a unified cancellation
mechanism.
One approach to solve this is to abandon task_mgr in favor of a
mechanism similar to Golang's context.Context, albeit extended to
support waiting for completion, and specialized to the needs in the
pageserver.

Heikki solves all of the above at once in PR
https://github.com/neondatabase/neon/pull/3228 , which is not yet
merged at the time of writing.

What Is This Patch About
========================

This patch addresses the immediate needs for layer eviction
by introducing a RequestContext structure (name up for debate)
that is plumbed through all the way from the entrypoints of the
pageserver (page_service, management API, tenant background loops)
all the way down to Timeline::{get,get_reconstruct_data}.

The struct carries a description of the kind of activity that initiated
the call (re-using task_mgr::TaskKind here).

Also, it carries the desired on-demand download behavior of the entrypoint.
Timeline::get_reconstruct_data can then log the TaskKind that initiated
the on-demand download.

I developed this patch by git-checking-out Heikki's big RequestContext
PR https://github.com/neondatabase/neon/pull/3228 , then deleting all
the functionality that we do not need to address the needs for layer
eviction.

This patch is huge, it should be easy to review because almost all
of it is plumbing through of the &RequestContext argument.

Also, this patch is low risk because it should not be changing
any behavior (apart from DownloadBehavior::Error).

TODO:
- Self-review
- We could reduce the size a bit by removing the various occurences of
  ```
  fn(..., _ctx: &RequestContext)
  ```
  that I left in place from PR 3228.