Add a wrapper to trace queries issued by 'compute_ctl'.
This allows more fine-grained tracing of how much time each query
takes during compute startup.

The tracing is done in a new wrapper crate around the 'postgres'
crate, so that it's as transparent as possible to the calling
code. The wrapper could be used elsewhere too, but for now it's only
used in 'compute_ctl'.