[1/X] remove TimelineState::Broken, simplify TimelineState transitions, internalize Timeline shutdown procedure
Before this patch, tenant.rs used the TimelineState::Broken to keep
track of initialization failure or success.
State Broken was not used for anything else.

After this patch, tenant.rs keeps track of that inside the
Tenant::timelines hash map, via the new TimelineMapEntry enum.

That enum also has a variant to indicate that a timeline is being
removed. The tenant.rs can now use this, instead of Timeline::current_state()
to check whether the timeline is removing or not.

All of the above unlocks:

1. race-free timeline deletion. Only one task can win the transition
   from TimelineMapEntry::Live to TimelineMapEntry::Removing.
2. Timeline::state is only changed from within timeline.rs
3. TimelineState::Broken is never written.

Given these insights, we can now push down the code that shuts
down the Timeline-scoped tasks into a Timeline::shutdown routine.
That routine can use the (now internal-to-struct-Timline)
Timeline::state to reason about what needs to be done when shutting
down a timeline.

The next step will be to eliminate TimelineState::Loading by
representing that state as a TimelineMapEntry variant instead.
Afterwards, the Timeline::state is effectively a CancellationToken,
which in turn will help the effort to unify cancellation and task management
in the pagserver.

This is PR: https://github.com/neondatabase/neon/pull/3449