WIP tenant size calculation: refactor, rewrite, and add SVG
Refactor the tenant_size_model code. Segment now contains just the
minimum amount of information needed to calculate the size. Other
information that is useful for building up the segment tree, and
for display purposes, is now kept elsewhere. The code in 'main.rs'
has a new ScenarioBuilder struct for that.

Calculating which Segments are "needed" is now the responsibility of
the caller of tenant_size_mode, not part of the calculation itself.
So it's up to the caller to make all the decisions with retention
periods for each branch.

The output of the sizing calculation is now a Vec of SizeResults,
rather than a tree. It uses a tree representation internally, when
doing the calculation, but it's not exposed to the caller anymore.

I also refactored the way the recursive calculation is performed. With
the previous coding, I wasn't quite sure what the complexity was,
O(n) or something worse. I'm now pretty convinced it's O(n), where 'n'
is the number of Segments.

Rewrite the code in size.rs that builds the Segment model. I got rid
of the intermediate representation with Update structs. It now builds
the Segments directly, with some local HashMaps and Vecs to track
branch points to help with that.

retention_period is now an input to gather_inputs(), rather than an
output.

Add 'size_debug' endpoint that displays an SVG file with the branches,
as well as all the information needed by the model as JSON.

TODOs:

- lots of misc cleanup

- I believe test_single_branch_get_tenant_size_grows was wrong before