add metrics to detect eviction-induced thrashing
This patch adds two metrics that will enable us to detect *thrashing* of
layers, i.e., repetitions of `eviction, on-demand-download, eviction,
... ` for a given layer.

The first metric counts all layer evictions per timeline.  It requires
no further explanation.  The second metric counts the layer evictions
where the layer was resident for less than a given threshold.

We can alert on increments to the second metric.  The first metric will
serve as a baseline, and further, it's generally interesting, outside of
thrashing.

The second metric's threshold is configurable in PageServerConf and
defaults to 24h.  The threshold value is reproduced as a label in the
metric because the counter's value is semantically tied to that
threshold.  Since changes to the config and hence the label value are
infrequent, this will have low storage overhead in the metrics storage.

The data source to determine the time that the layer was resident is the
file's `mtime`.  Using `mtime` is more of a crutch.  It would be better
if Pageserver did its own persistent bookkeeping of residence change
events instead of relying on the filesystem.  We had some discussion
about this:
https://github.com/neondatabase/neon/pull/3809#issuecomment-1470448900

My position is that `mtime` is good enough for now.  It can
theoretically jump forward if someone copies files without resetting
`mtime`.  But that shouldn't happen in practice.  Note that moving files
back and forth doesn't change `mtime`, nor does `chown` or `chmod`.
Lastly, `rsync -a`, which is typically used for filesystem-level backup
/ restore, correctly syncs `mtime`.

I've added a label that identifies the data source to keep options open
for a future, better data source than `mtime`.  Since this value will
stay the same for the time being, it's not a problem for metrics
storage.

refs https://github.com/neondatabase/neon/issues/3728