refactor: abstraction for the below-threshold counter
This wraps the following aspects of the
evictions_with_low_residence_duration counter into an abstraction:

* observations are only counted if below threshold
  * Previously, this was done at the call site
* Adding the data source and threshold label values
* Removal of the metric values

It's a shame that the underlying `prometheus` library doesn't allow
currying of the labels. We could turn this into a more generic
abstraction otherwise.

Is this refactor a net win?
Pro: yes because it encapsulates above aspects.
No: not in line with how other metrics are done, they don't have their
    own wrapper type