hack: avoid redownloads post-restart by synthetic_size
with #3867 we started to retain a set of layers which will be needed
right after the next restart. with this hack, we will also retain the
layers which are needed most likely for the next synthetic size
calculation.

the actual size calculation is not of importance, but we need to
calculate all of the sizes, so we only call tenant::size::gather_inputs.

Co-authored-by: Christian Schwarz <christian@neon.tech>