random_init_delay: remove the minimum of 10 seconds
Before this patch, the range from which the random delay is picked
is at minimum 10 seconds.
With this patch, they delay is bounded to whatever the given
`period` is, and zero, if period id Duration::ZERO.

Motivation for this: the disk usage eviction tests that we'll add in
https://github.com/neondatabase/neon/pull/3905 need to wait for the disk
usage eviction background loop to do its job.
They set a period of 1s.
It seems wasteful to wait 10 seconds in the tests.

Co-authored-by: Joonas Koivunen <joonas@neon.tech>