[compute_ctl] Improve 'empty' compute startup sequence
Do several attempts to get spec from the control-plane and retry network
errors and all reasonable HTTP response codes. Do not hang waiting for
spec without confirmation from the control-plane that compute is known
and is in the `Empty` state.

Adjust the way we track `total_startup_ms` metric, it should be
calculated since the moment we received spec, not from the moment
`compute_ctl` was started. Also introduce a new `wait_for_spec_ms`
metric to track the time spent sleeping and waiting for spec to be
delivered from control-plane.

Part of neondatabase/cloud#3533