Improved storage of records in the delta layer btree
Before, we stored each WAL record separately in the main index. With this
change, the main btree only contains:

 - References to Images, and WAL records that apply on top of them
 - References to sequences of WAL records that apply to the same page, but
   don't have an image in this Layer

This reduces the size of the index by some amount, and thus increases the cache-ability of that index.

Unless we're not looking for the latest version of a page or otherwise limit
the lookup window, this change does not significantly impact IO requirements
for normal workloads, as we don't (yet) add any compression to this WAL.